requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.7
django-import-export==4.1.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1